+++ CHANGES.rst
@@ -2,7 +2,62 @@ Changelog
 =========
 
 
-1.0a3 (unreleased)
+1.0b5 (unreleased)
+------------------
+
+- Avoid `UnicodeDecodeError` in `SignersColumn` if signers
+  contain mixed encoding values.
+  [gbastien]
+- Added session info size on quick look.
+  [chris-adam]
+
+1.0b4 (2026-03-27)
+------------------
+
+- Renamed imio.esign config functions.
+  [cadam]
+- Highlight `draft` session in table view and viewlet, use `Id` as
+  column header instead `identifier` so it is more narrow and
+  make `sessions` view columns sortable.
+  [gbastien]
+
+1.0b3 (2026-03-26)
+------------------
+
+- In sessions annotation, use a `PersistentMapping` instead a `dict` to store
+  `signers` informations as the `status` will be updated.
+  [gbastien]
+
+1.0b2 (2026-03-26)
+------------------
+
+- Redo release.
+  [gbastien]
+
+1.0b1 (2026-03-26)
+------------------
+
+- Disable sorting on checkbox column of `@@signing-users-csv`.
+  [gbastien]
+- Added SessionAnnotationInfoView.
+  [chris-adam,sgeulette]
+- Open the `@@session-annotation-info` in an overlay.
+  [gbastien]
+- Fixed audit message when firing action remove context
+  [chris-adam]
+
+1.0a4 (2026-03-24)
+------------------
+
+- Redirect to `HTTP_REFERER` instead context's url after
+  `add/remove/remove item` actions.
+  [gbastien]
+- Added `odd/even` class on `<li>` of the files `collapsible` on sessions view.
+  [gbastien]
+- Added separated fingerpointing log.
+  [chris-adam]
+
+1.0a3 (2026-03-20)
 ------------------
 
 - Replaced external_session_link p by span.
@@ -29,8 +84,20 @@ Changelog
   [chris-adam]
 - Added external watchers for esign sessions.
   [chris-adam, sgeulette]
-- Added session info size on quick look.
+- Added possibility to have elements of the same context to belong to different sessions.
   [chris-adam]
+- Manage file added again to same session, data is updated.
+  [gbastien]
+- Configured the `@@remove-from-esign-session` the same way as `@@remove-item-from-esign-session`
+  so relying on an `available` method to show it only if a context is in a session.
+  [gbastien]
+- Fixed faceted viewlet broken because sessions format changed from list to OrderedDict.
+  [gbastien]
+- Turned file info in session annotation from `dict` to `PersistentMapping`.
+  Entire annotation structure is now persistent, remove `_p_changed`.
+  [gbastien]
+- Use `@CachedProperty` for `FacetedSessionInfoViewlet.sessions` and `ItemSessionInfoViewlet.sessions`.
+  [gbastien]
 
 1.0a2 (2026-02-06)
 ------------------